src/requirements.txt
BeautifulSoup==3.2.1
Django==1.5.4
GitPython==0.3.2.RC1
Markdown==2.1.0
Pygments==1.4
Whoosh==2.5.6
argparse==1.2.1
async==0.6.1
beautifulsoup4==4.3.2
django-appconf==0.6
django-haystack==2.1.0
gitdb==0.5.4
markdown2==2.2.0
pytz==2011n
repoze.timeago==0.5
six==1.5.2
smmap==0.8.1
wsgiref==0.1.2